myapp/requirements.txt
firebase-admin==6.4.0
Flask==3.0.2
Flask-Cors==4.0.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-secret-manager==2.17.0
google-cloud-storage==2.14.0
gunicorn==21.2.0
openai==1.12.0
python-dotenv==1.0.1
httpx==0.27.0
